requirements.txt
asgiref==3.3.0
boto3==1.16.13
botocore==1.19.13
Django==3.1.3
django-environ==0.4.5
django-storages==1.10.1
gunicorn==20.0.4
jmespath==0.10.0
python-dateutil==2.8.1
pytz==2020.4
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.11
